move driver to cmip6 module
